fixed all fixable warnings about implicit type conversion

+++ Minuit/inc/Minuit2/FumiliMaximumLikelihoodFCN.h
@@ -144,14 +144,14 @@ public:
 
     double sumoflogs = 0.0;
     std::vector<double> vecElements =  Elements(par);
-    unsigned int vecElementsSize = vecElements.size();
+    unsigned vecElementsSize = static_cast<unsigned>(vecElements.size());
 
-    for (unsigned int i = 0; i < vecElementsSize; ++i) {
+    for (unsigned i = 0; i < vecElementsSize; ++i) {
       double tmp = vecElements[i];
       //for max likelihood probability have to be positive
       assert(tmp >= 0);
       if ( tmp < FLT_MIN*5 )
-	tmp = FLT_MIN*5;
+        tmp = FLT_MIN*5;
 
       sumoflogs -= std::log(tmp);
       //std::cout << " i " << tmp << " lik " << sumoflogs << std::endl;
